fastapi-projects/15-netra-vision/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
httpx
Pillow
python-multipart
